Add MTD configuration for operational dataset management. (#999)
